pyproject.toml
[project]
name = "diecast"
version = "0.0.0"
description = "Diecast — open-source extraction of TaskOS / LinkedOut agent harness."
requires-python = ">=3.11"
readme = "README.md"
license = { text = "Apache-2.0" }
dependencies = [
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "jsonschema>=4.26.0",
    "psutil>=5.9",
    "pytest>=8.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["cast-server"]
include = ["cast_server*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra"
markers = [
    "integration: integration test (spawns real processes / binds ports)",
]
